sensor data backend have done !

+++ app/src/main/java/com/github/oldr1990/model/MappedBMEData.kt
@@ -3,6 +3,6 @@ package com.github.oldr1990.model
 data class MappedBMEData(
     val humidity: Float,
     val temperature: Float,
-    val pressure: Float,
+    val pressure: Int,
     val date: String
 )
